Move model classes into separate namespace
