+++ NextDepartures.Standard/GetServicesByStopAsync.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using NextDepartures.Standard.Model;
+using System;
 using System.Collections.Generic;
 using System.Data;
 using System.Data.SqlClient;
