+++ NextDepartures.Standard/Model/Agency.cs
@@ -1,0 +1,9 @@
+﻿namespace NextDepartures.Standard.Model
+{
+    internal class Agency
+    {
+        public string AgencyID { get; set; }
+        public string AgencyName { get; set; }
+        public string AgencyTimezone { get; set; }
+    }
+}
