+++ NextDepartures.Standard/Model/Departure.cs
@@ -1,0 +1,24 @@
+﻿namespace NextDepartures.Standard.Model
+{
+    internal class Departure
+    {
+        public string AgencyID { get; set; }
+        public string DepartureTime { get; set; }
+        public string EndDate { get; set; }
+        public string Friday { get; set; }
+        public string Monday { get; set; }
+        public string RouteLongName { get; set; }
+        public string RouteShortName { get; set; }
+        public string Saturday { get; set; }
+        public string ServiceID { get; set; }
+        public string StartDate { get; set; }
+        public string StopID { get; set; }
+        public string Sunday { get; set; }
+        public string Thursday { get; set; }
+        public string TripHeadsign { get; set; }
+        public string TripID { get; set; }
+        public string TripShortName { get; set; }
+        public string Tuesday { get; set; }
+        public string Wednesday { get; set; }
+    }
+}
