+++ NextDepartures.Standard/Model/Exception.cs
@@ -1,0 +1,9 @@
+﻿namespace NextDepartures.Standard.Model
+{
+    internal class Exception
+    {
+        public string Date { get; set; }
+        public string ExceptionType { get; set; }
+        public string ServiceID { get; set; }
+    }
+}
