+++ NextDepartures.Standard/Model/Service.cs
@@ -1,0 +1,12 @@
+﻿namespace NextDepartures.Standard.Model
+{
+    public class Service
+    {
+        public string AgencyName { get; set; }
+        public string DepartureTime { get; set; }
+        public string DestinationName { get; set; }
+        public string RouteName { get; set; }
+        public string StopName { get; set; }
+        public string TripID { get; set; }
+    }
+}
