+++ NextDepartures.Standard/Model/Stop.cs
@@ -1,0 +1,9 @@
+﻿namespace NextDepartures.Standard.Model
+{
+    public class Stop
+    {
+        public string StopID { get; set; }
+        public string StopName { get; set; }
+        public string StopTimezone { get; set; }
+    }
+}
